More protocol-agnostic, nested proxies
This no longer depends on the root type having "name" and "params"
fields; it should work nicely with any given shallow type.

To allow arbitrary shallowTypes, we recursively create proxies for every
field on the packet. `makeProxyHandler(path)` returns a closure which
captures `path` and lets each proxy keep track of where the proxied
object is in the packet, so if the full packet data needs to be parsed,
it can access the correct (possibly deeply nested) field in the
fully-parsed packet.

For example, if shallow-parsing some packet yields the structure:
{
    name: 'packet_llama_panda',
    params: {
	llama: {}
    }
}
and if fully parsing it yields the structure:
{
    name: 'packet_llama_panda',
    params: {
	llama: {
	    panda: 137
	}
    }
}
then if we do:
```
const llamaProxy = packet.params.llama // doesn't parse the full packet yet
const panda = llamaProxy.panda // panda is not found in the shallow structure, so the packet gets fully parsed
```
then `panda` will get the correct value, 137. If we didn't keep track of
the access path, then llamaProxy would have no way to know where to look for
`panda` in the full packet data.

I'm not sure whether this strategy will yield the best performance, but
I couldn't think of any great alternatives. Another option would be to
re-use the same proxy handler for nested proxies, removing the need for
the closure, and do some kind of deep copy into the shallow packet once
the full packet is parsed. Or we could just only support shallowTypes
with a shallow, opinionated structure like {name, params: {}}.

In my testing, the lazy parser was around 10 times slower than
FullPacketParser in NMP's benchmark.js. But we should definitely try
comparing them in a real-world environment where not every packet needs
to be parsed.