Benchmarking full vs. lazy parsing
Still WIP. Depends on test/packet.json in evan-goode/ProtoDef#lazy.
Right now, that file just contains a copy of the u8/u32 container from
conditional.json, but eventually it should have several large, complex
types that are representative of packets you'd find in the real world.

Not sure if this is a good way to structure this, we might want a
separate schema for defining parser tests versus datatype tests?